Multiplatform Makefile
Since I decided to kinda support windows, I needed a way to compile different
targets. The best way I knew was using different config.mk files that assign
different dependecies, targets and so on.

That kind of approach requires the Makefile to be POSIX-complaint if the
project is to be compiled on something else than Linux or Windows. I used
pdpmake for that, althogh I had to use 'target_name' pragma so that targets
could have slashes in them. So the Makefile is not 100% POSIX, but I think
most platform's make impelementations have that feature.

The tags idea like +linux and +mingw was taken from Drew DeVault. I don't
know if I'm doing it properly, though.